core/app/crash/reporting: Declare ReportMinidump stub.
If the crashreporting build tag isn't defined, we fallback to a noop stub implementation. This stub was missing the new `ReportMinidump` function, breaking builds that didn't have the crashreporting tag defined.
